Rm unnecessary resources, reactor RockImages
[#184027461]
